fix regression tests, all but one
the one that's "failing" is actually failing because the algorithm
successfully computed inputs, and the test asserts that it can't. but
that's kind of suspect - it's seems possible that I was never able to
get the old algorithm to compute a result in the first place so that I
could verify it, and I just assumed it was correct and had the test
assert against it. (all regression tests are performance, where we
initially do not expect a particular result, only that it succeeds at
all.)

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>