start containerd ourselves
it seems important to ensure the containerd client library is in sync
with the containerd server. bundling containerd with concourse, the same
way docker does, is one way to ensure this.

in addition, pointing to an arbitrary containerd socket won't exactly
work anyway, since the client process has to be local to the containerd
machine in order to manipulate snapshots.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>