atc: structure: BuildFactory -> BuildPlanner
...and rewrite the test suite.

I found myself spending way too much time examining the tests in the
package whenever I made a seemingly simple change to the Plan.

This package is largely a mechanical task of making sure the right
PlanConfig turns into the write Plan, so it's a good fit for
table-driven tests. This should hopefully be much lower maintenance.

The tests still assert that plan IDs are unique, and assert on exact
plan IDs when necessary. Otherwise they are omitted to reduce noise and
make the tests a bit easier to write.

The precedence between attempts:, timeout:, and the various hooks is
also covered.

Signed-off-by: Alex Suraci <asuraci@pivotal.io>