atc/gc: structure: add EventStore with Delete method
An `EventStore` is a build processor that'll have some extra
functionality. This commit adds the `Delete` method, and uses it in the
build log collector.

We'll also need to call this explicitly when deleting Pipelines and
Teams. Currently, that's done using Postgres triggers (likewise with the
creation of the team_build_events and pipeline_build_events tables). If
we'll want to fully decouple event processing from the database, we'll
need to get rid of these triggers in migrations.

If we were to remove the triggers in migrations, we'd need some way of
either adding them back, or have a new mechanism for creating and
deleting these tables.

If we wanted to add the triggers in code, likely the best approach would
be to have an idempotent `Init` function that gets called on startup for
the chosen `EventStore`. This seems like a good idea anyway, because
we'll usually want to setup some connection to a remote resource once
initially (this isn't a problem for the DB event store since we already
set up a connection, which we can just pass in)

Otherwise, we could create the table if it doesn't exist when calling
Initialize(Build) on the first build in a pipeline/team. When deleting a
list of build IDs, we could check if there are no longer any builds in
the table, and if so, delete the table.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>