atc: behaviour: remove Last-Event-ID header
This header allows users to query the /api/v1/builds/{id}/events
endpoint starting from an offset.

Removing this may be a controversial decision. I discussed this as an
open question in concourse/rfcs#53. Essentially, supporting this header
means that `EventStores` will have to define how to unmarshal `Keys`
(but not marshal them??).

This header seems to be used nowhere internally, and isn't exposed by
Fly or go-concourse, so my guess is it's *probably* no longer used by anyone.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>