atc: behaviour: propagate context to EventStore
This also involves some of the db methods taking in a context, which has
pretty sweeping code changes, but is probably a good idea (especially
since we can execute all queries with a context). Eventually, it could
be cool to have all of our db methods take in a context - this would let
us add tracing. This may be overkill, since Postgres already takes
statistics on query execution times, but still - could be good for
observability.

P.S. this broke a lot of the test code - test changes to follow in a
later commit. It should hopefully be pretty skimmable.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>