atc: behaviour: use notification bus for event stream
Making use of the updated notification bus implementation from the prior
commit, use the use the notification bus for receiving events (rather
than Getting from the EventStore every time there's a new event). Now,
we call `EventStore.Get` initially to get the events that are already
stored, and use the notification bus for new events.

Recall, the motivation for this is to get around a limitation with
Elasticsearch where storing documents in an index does not immediately
allow them to be searchable (by default - and getting around this can
lead to poor performance). I imagine it has some performance
implications. I'm not sure if it'll be an improvement to performance or
a hindrance - it definitely requires some more investigation.

This change also requires adding a total ordering to Keys, which I think
is nice (rather than having them as `interface{}`). It also requires
`Marshal` and `Unmarshal` behaviour - This means we can bring back the
`Last-Event-ID` header if there's need for it.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>