db: structure: rename Key -> EventKey
`db.Key` doesn't really tell you what it is. Ideally, I would be able to
move these interfaces elsewhere, but that's not so easy do to import
cycles (db needs to import the package that houses the EventStore
interface, and the EventStore signature includes db types).

I thought about moving `Key` to the `event` package - this way, it's an
`event.Key` which is nice - but felt it was better to keep `EventKey`
and `EventStore` as close to each other as possible (since `EventKey` is
only useful in the presence of its corresponding `EventStore`, anyway)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>