db: structure: remove goto in eventsource
To be honest, I feel that using goto here is a clearer way to write the
code. It's inherently *not* a loop (restarting is only for the errant
case), so having a loop front-and-center makes the code a bit harder to
grok IMO (and really, `continue` in this context *is* just `goto start`)

That said, it's the only use of goto in the codebase, so I'd rather not
be too controversial.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>