web: structure: group pipelines by team name
Store pipelines grouped by team name rather than in a flat list to
facilitate reordering them. Previously, reordering (i.e. drag and drop)
relied on the fact that all of a team's pipelines were stored adjacent
to one another in the flat list. This has the added benefit of likely
improved performance when modifying a pipeline/filtering, though that
isn't the primary motivation.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>