web: behaviour: fix pipeline reordering with hidden pipelines
This appears to be a structural change, and if a `Group`'s `.pipelines`
were the same as the full list of that team's pipelines (i.e. the team
has no archived pipelines, meaning all pipelines are displayed), there
is no behaviour change, and thinking in terms of `DropTargets` rather
than indexes is simply a more intuitive way of thinking about Drag and
Drop.

When `Group.pipelines` does not reflect all of the team's pipelines (for
instance, the team has archived pipelines which are therefore not
displayed), `DropTargets` relative to pipeline names still work (unlike
indexes) - e.g. if `pipeline1` is the name of an unarchived pipeline,
`Before "pipeline1"` is always a valid place to drop a pipeline - it may
have a different index in `Dashboard.Model.pipelines` than it does in
`Group.pipelines` (since other archived pipelines appear before it).

All that to say, pipeline reordering now works even when there are
pipelines that aren't rendered

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>