web: behaviour: generate event ids with postgres
Previously, Elasticsearch used the timestamp of the build events + a
"tiebreaker" (an offset per-ATC counter) to uniquely identify events.
The main problem with this is that different ATCs may not agree on the
timestamp, so newer events may appear older (and therefore not be sent
to the user/appear out of order) if they come from different ATCs.

To remedy this problem, make use of Postgres to generate auto
incrementing IDs for the Elasticsearch backend.

This commit also forces the `EventKey` to be a uint, which makes several
things easier - no need for special marshal/unmarshal logic, can be used
with external plugin implementations, etc. Any backend that does not
support auto incrementing natively (like Elasticsearch) can just use
Postgres to generate the IDs.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>